refactor: restructure URL hierarchy
