+++ spog/ui/src/pages/mod.rs
@@ -3,7 +3,6 @@
 use yew_nested_router::Target;
 
 mod advisory;
-mod catalog;
 mod chicken;
 mod index;
 mod package;
@@ -11,11 +10,9 @@ mod sbom;
 // mod vulnerability;
 
 pub use advisory::*;
-pub use catalog::*;
 pub use chicken::*;
 pub use index::*;
 pub use package::Package;
-pub use package::*;
 pub use sbom::*;
 // pub use vulnerability::Vulnerability;
 
@@ -24,19 +21,25 @@ pub enum AppRoute {
     #[target(index)]
     Index,
     Chicken,
-    Package {
-        query: String,
-    },
-    Catalog {
-        query: String,
-    },
-    SBOM {
-        id: String,
-    },
+    Package(View),
     Advisory {
         query: String,
     },
     /*Vulnerability {
         query: String,
     },*/
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Target)]
+pub enum View {
+    Search { query: String },
+    Content { id: String },
+}
+
+impl Default for View {
+    fn default() -> Self {
+        Self::Search {
+            query: Default::default(),
+        }
+    }
 }
